backend/requirements.txt
bottle==0.13.4
cffi==2.0.0
clr_loader==0.2.10
imageio-ffmpeg==0.6.0
proxy-tools==0.1.0
pycparser==3.0
pythonnet==3.0.5
pywebview==6.1
typing_extensions==4.15.0
yt-dlp==2026.2.21
